requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.3.0
albumentations>=1.3.0

# Data handling
pandas>=1.3.0
scikit-learn>=1.0.0
h5py>=3.1.0

# Configuration and utilities
PyYAML>=6.0
omegaconf>=2.1.0
hydra-core>=1.1.0
python-dotenv>=0.19.0

# Logging and monitoring
tensorboard>=2.8.0
wandb>=0.13.0
tqdm>=4.62.0

# Model architectures and pre-trained models
timm>=0.9.0
segmentation-models-pytorch>=0.3.0
torchmetrics>=0.11.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Development tools
jupyter>=1.0.0
ipython>=7.25.0
notebook>=6.4.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Code quality
black>=21.6.0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910

# Additional utilities
click>=8.0.0
rich>=10.0.0
typer>=0.4.0